CTCP-4540: move files around to match ken structure
