+++ app/pages/departureMeansOfTransport/VehicleIdentificationNumberPage.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package pages
+package pages.departureMeansOfTransport
 
 import models.Index
+import pages.QuestionPage
 import pages.sections.TransportMeansListSection
 import play.api.libs.json.JsPath
