+++ conf/app.departureMeansOfTransport.routes
@@ -2,3 +2,9 @@ GET        /:arrivalId/departure-means-of-transport/:transportMeansIndex/country
 POST       /:arrivalId/departure-means-of-transport/:transportMeansIndex/country                                     controllers.departureMeansOfTransport.CountryController.onSubmit(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = NormalMode)
 GET        /:arrivalId/departure-means-of-transport/:transportMeansIndex/country                                     controllers.departureMeansOfTransport.CountryController.onPageLoad(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = CheckMode)
 POST       /:arrivalId/departure-means-of-transport/:transportMeansIndex/country                                     controllers.departureMeansOfTransport.CountryController.onSubmit(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = CheckMode)
+
+
+GET        /:arrivalId/departure-means-of-transport/:transportMeansIndex/identification-number                       controllers.departureMeansOfTransport.VehicleIdentificationNumberController.onPageLoad(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = NormalMode)
+POST       /:arrivalId/departure-means-of-transport/:transportMeansIndex/identification-number                       controllers.departureMeansOfTransport.VehicleIdentificationNumberController.onSubmit(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = NormalMode)
+GET        /:arrivalId/departure-means-of-transport/:transportMeansIndex/change-identification-number                controllers.departureMeansOfTransport.VehicleIdentificationNumberController.onPageLoad(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = CheckMode)
+POST       /:arrivalId/departure-means-of-transport/:transportMeansIndex/change-identification-number                controllers.departureMeansOfTransport.VehicleIdentificationNumberController.onSubmit(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = CheckMode)
