+++ conf/app.routes
@@ -31,12 +31,6 @@ POST       /:arrivalId/goods-unloaded-date                                      
 GET        /:arrivalId/change-goods-unloaded-date                                                                    controllers.DateGoodsUnloadedController.onPageLoad(arrivalId: ArrivalId, mode: Mode = CheckMode)
 POST       /:arrivalId/change-goods-unloaded-date                                                                    controllers.DateGoodsUnloadedController.onSubmit(arrivalId: ArrivalId, mode: Mode = CheckMode)
 
-GET        /:arrivalId/departure-means-of-transport/:transportMeansIndex/identification-number                            controllers.VehicleIdentificationNumberController.onPageLoad(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = NormalMode)
-POST       /:arrivalId/departure-means-of-transport/:transportMeansIndex/identification-number                            controllers.VehicleIdentificationNumberController.onSubmit(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = NormalMode)
-GET        /:arrivalId/departure-means-of-transport/:transportMeansIndex/change-identification-number                     controllers.VehicleIdentificationNumberController.onPageLoad(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = CheckMode)
-POST       /:arrivalId/departure-means-of-transport/:transportMeansIndex/change-identification-number                     controllers.VehicleIdentificationNumberController.onSubmit(arrivalId: ArrivalId, transportMeansIndex: Index, mode: Mode = CheckMode)
-
-
 GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/gross-weight                        controllers.GrossWeightController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = NormalMode)
 POST       /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/gross-weight                        controllers.GrossWeightController.onSubmit(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = NormalMode)
 GET        /:arrivalId/house-consignment/:houseConsignmentIndex/items/:itemIndex/change-gross-weight                 controllers.GrossWeightController.onPageLoad(arrivalId: ArrivalId, houseConsignmentIndex: Index, itemIndex: Index, mode: Mode = CheckMode)
